+++ ropwr/base.py
@@ -24,8 +24,9 @@ from .cvx_qp import qp_separated
 
 
 def _check_parameters(objective, regularization, degree, continuous,
-                      monotonic_trend, solver, h_epsilon, quantile, reg_l1,
-                      reg_l2, verbose):
+                      continuous_deriv, monotonic_trend, solver, h_epsilon,
+                      quantile, reg_l1, reg_l2, max_iter, extrapolation,
+                      extrapolation_bounds, space, verbose):
 
     if objective not in ("l1", "l2", "huber", "quantile"):
         raise ValueError('Invalid value for objective. Allowed string '
@@ -41,7 +42,11 @@ def _check_parameters(objective, regularization, degree, continuous,
 
     if not isinstance(continuous, bool):
         raise TypeError("continuous must be a boolean; got {}."
-                        .format(verbose))
+                        .format(continuous))
+
+    if not isinstance(continuous_deriv, bool):
+        raise TypeError("continuous_deriv must be a boolean; got {}."
+                        .format(continuous_deriv))
 
     if monotonic_trend is not None:
         if monotonic_trend not in ("ascending", "descending", "convex",
@@ -75,6 +80,31 @@ def _check_parameters(objective, regularization, degree, continuous,
     if not isinstance(reg_l2, numbers.Number) or reg_l2 < 0.0:
         raise ValueError("reg_l2 must be a positive value; got {}."
                          .format(reg_l2))
+
+    if max_iter is not None:
+        if not isinstance(max_iter, numbers.Integral) or max_iter <= 0:
+            raise ValueError("max_iter must be an integer >= 1; got {}."
+                             .format(max_iter))
+
+    if extrapolation is not None:
+        if extrapolation not in ("constant", "continue", "linear"):
+            raise ValueError('Invalid value for extrapolation. Allowed string '
+                             'values are "constant", "continue" and "linear".')
+
+    if extrapolation_bounds is not None:
+        if not isinstance(extrapolation_bounds, tuple):
+            raise TypeError("extrapolation_bounds must be a tuple or None.")
+
+        if len(extrapolation_bounds) != 2:
+            raise ValueError("extrapolation_bounds tuple must contain two "
+                             "elements (lower_bound, upper_bound).")
+
+    if space not in ("linear", "log"):
+        raise ValueError('Invalid value for space. Allowed string values are '
+                         '"linear" and "log".')
+
+    if space == "log" and extrapolation == "linear":
+        raise ValueError("extrapolation cannot be linear if log space.")
 
     if not isinstance(verbose, bool):
         raise TypeError("verbose must be a boolean; got {}.".format(verbose))
@@ -202,6 +232,14 @@ def _check_splits(x, splits, n_bins, monotonic_trend):
             sorted_idx = np.argsort(user_splits)
             user_splits = user_splits[sorted_idx]
 
+            # Check if any split is exactly 0. This create issues when
+            # the degree of the polynomial > 1. Add epsilon. In general,
+            # cast to float to avoid power(integer, neg_exponent) error
+            user_splits = user_splits.astype(float)
+
+            if np.any(user_splits == 0):
+                user_splits[user_splits == 0] += 1e-6
+
         return user_splits
 
 
@@ -228,6 +266,11 @@ class RobustPWRegression(BaseEstimator):
 
     continuous : bool (default=True)
         Whether to fit a continuous or discontinuous piecewise regression.
+
+    continuous_deriv : bool (default=True)
+        Whether to fit a polynomial with continuous derivatives. This option
+        fits a smooth degree d-polynomial with d-1 continuity in derivatives
+        (splines).
 
     monotonic_trend : str or None, optional (default=None)
         The monotonic trend. Supported trends are "ascending", "descending",
@@ -260,6 +303,35 @@ class RobustPWRegression(BaseEstimator):
         L2 regularization term. Increasing this value will smooth the
         regression model. Only applicable if ``regularization="l2"``.
 
+    max_iter : int or None (default=None)
+        Maximum number of iterations. If ``max_iter=None``, the default maximum
+        number of iterations for each solver is set. See `solver options
+        <https://www.cvxpy.org/tutorial/advanced/index.html#
+        setting-solver-options>`_.
+
+        .. versionadded:: 1.0.0
+
+    extrapolation : str or None (default="constant")
+        If None, an exception is raised if values outside the range (min, max)
+        of the fitting. If 'constant', the value of the regression at min, max
+        value of the fitting is used for constant extrapolation. If 'continue',
+        the regression is extrapolated as is. If 'linear', a linear
+        extrapolation is used.
+
+        .. versionadded:: 1.0.0
+
+    extrapolation_bounds : tuple or None (default=None)
+        If tuple ``(lb, ub)``, the regression prediction is limited to values
+        within bounds. If None, the regression prediction is not limited.
+
+        .. versionadded:: 1.0.0
+
+    space : str, optional (default="linear")
+        The y space. If ``space="log"``, a logarithm transformation is applied
+        to the input data y.
+
+        .. versionadded:: 1.0.0
+
     verbose : bool (default=False)
         Enable verbose output.
 
@@ -269,13 +341,15 @@ class RobustPWRegression(BaseEstimator):
         Coefficients for each bin. Number of bins = n_splits + 1.
     """
     def __init__(self, objective="l2", degree=1, continuous=True,
-                 monotonic_trend=None, solver="auto", h_epsilon=1.35,
-                 quantile=0.5, regularization=None, reg_l1=1.0, reg_l2=1.0,
-                 verbose=False):
+                 continuous_deriv=True, monotonic_trend=None, solver="auto",
+                 h_epsilon=1.35, quantile=0.5, regularization=None, reg_l1=1.0,
+                 reg_l2=1.0, max_iter=None, extrapolation="constant",
+                 extrapolation_bounds=None, space="linear", verbose=False):
 
         self.objective = objective
         self.degree = degree
         self.continuous = continuous
+        self.continuous_deriv = continuous_deriv
         self.monotonic_trend = monotonic_trend
         self.solver = solver
         self.h_epsilon = h_epsilon
@@ -283,9 +357,16 @@ class RobustPWRegression(BaseEstimator):
         self.regularization = regularization
         self.reg_l1 = reg_l1
         self.reg_l2 = reg_l2
+        self.max_iter = max_iter
+        self.extrapolation = extrapolation
+        self.extrapolation_bounds = extrapolation_bounds
+        self.space = space
         self.verbose = verbose
 
         self.coef_ = None
+
+        self._min_x = None
+        self._max_x = None
 
         self._is_fitted = False
 
@@ -308,6 +389,8 @@ class RobustPWRegression(BaseEstimator):
             The number of bins to produce. Only applicable if splits is
             'uniform' or 'quantile'.
 
+            .. versionadded:: 0.4.0
+
         lb : float or None (default=None)
             Add constraints to avoid values below the lower bound lb.
 
@@ -326,9 +409,18 @@ class RobustPWRegression(BaseEstimator):
         y = check_array(y, ensure_2d=False, force_all_finite=True)
         check_consistent_length(x, y)
 
+        if self.space == "log":
+            if np.any(y <= 0):
+                raise ValueError("y data must be > 0 to use log space.")
+            y = np.log(y)
+
         idx = np.argsort(x)
         xs = x[idx]
         ys = y[idx]
+
+        # Store min and max x for extrapolation
+        self._min_x = xs[0]
+        self._max_x = xs[-1]
 
         # Check user splits
         splits = _check_splits(xs, splits, n_bins, self.monotonic_trend)
@@ -339,31 +431,39 @@ class RobustPWRegression(BaseEstimator):
         if bounded:
             _check_bounds(lb, ub)
 
+            if self.space == "log":
+                lb = np.log(lb) if lb is not None else None
+                ub = np.log(ub) if ub is not None else None
+
         # Choose the most appropriate method/solver given the parameters
         _method = _choose_method(self.objective, self.degree, self.continuous,
                                  self.monotonic_trend, self.solver, bounded,
                                  self.regularization)
 
         if _method == "lsq_direct":
-            c, info = lsq_direct(xs, ys, splits, self.degree)
+            c, info = lsq_direct(xs, ys, splits, self.degree,
+                                 self.continuous_deriv)
         elif _method == "lsq_direct_separated":
             c, info = lsq_direct_separated(xs, ys, splits, self.degree)
         elif _method == "socp":
-            c, info = socp(xs, ys, splits, self.degree, self.continuous, lb,
-                           ub, self.objective, self.monotonic_trend,
-                           self.h_epsilon, self.quantile, self.regularization,
-                           self.reg_l1, self.reg_l2, self.solver, self.verbose)
+            c, info = socp(xs, ys, splits, self.degree, self.continuous,
+                           self.continuous_deriv, lb, ub, self.objective,
+                           self.monotonic_trend, self.h_epsilon, self.quantile,
+                           self.regularization, self.reg_l1, self.reg_l2,
+                           self.solver, self.max_iter, self.verbose)
         elif _method == "socp_separated":
             c, info = socp_separated(xs, ys, splits, self.degree, lb, ub,
                                      self.objective, self.monotonic_trend,
                                      self.h_epsilon, self.quantile,
-                                     self.solver, self.verbose)
+                                     self.solver, self.max_iter, self.verbose)
         elif _method == "qp":
-            c, info = qp(xs, ys, splits, self.degree, self.continuous, lb, ub,
-                         self.monotonic_trend, self.verbose)
+            c, info = qp(xs, ys, splits, self.degree, self.continuous,
+                         self.continuous_deriv, lb, ub, self.monotonic_trend,
+                         self.max_iter, self.verbose)
         elif _method == "qp_separated":
             c, info = qp_separated(xs, ys, splits, self.degree, lb, ub,
-                                   self.monotonic_trend, self.verbose)
+                                   self.monotonic_trend, self.max_iter,
+                                   self.verbose)
 
         self.coef_ = c
 
@@ -414,21 +514,15 @@ class RobustPWRegression(BaseEstimator):
         p : numpy array, shape = (n_samples,)
             Predicted array.
         """
-        return self.fit(x, y, splits, n_bins, lb, ub).predict(x, lb, ub)
+        return self.fit(x, y, splits, n_bins, lb, ub).predict(x)
 
-    def predict(self, x, lb=None, ub=None):
+    def predict(self, x):
         """Predict using the piecewise regression.
 
         Parameters
         ----------
         x : array-like, shape = (n_samples,)
             Training vector, where n_samples is the number of samples.
-
-        lb : float or None (default=None)
-            Values below the lower bound lb are clipped to lb.
-
-        ub : float or None (default=None)
-            Values above the upper bound ub are clipped to ub.
 
         Returns
         -------
@@ -437,9 +531,20 @@ class RobustPWRegression(BaseEstimator):
         """
         self._check_is_fitted()
 
-        bounded = (lb is not None or ub is not None)
+        y_min, y_max = None, None
+        lb, ub = -np.inf, np.inf
 
-        if bounded:
+        if self.extrapolation is None:
+            if np.any((x < self._min_x) | (x > self._max_x)):
+                raise ValueError("x outside interpolation range.")
+
+        if self.extrapolation in ("constant", "linear"):
+            # Calculate predict(min_x) and predict(max_x)
+            y_min = np.polyval(self.coef_[0, :][::-1], self._min_x)
+            y_max = np.polyval(self.coef_[-1, :][::-1], self._max_x)
+
+        if self.extrapolation_bounds is not None:
+            lb, ub = self.extrapolation_bounds
             _check_bounds(lb, ub)
 
         n_bins = len(self._splits) + 1
@@ -450,8 +555,28 @@ class RobustPWRegression(BaseEstimator):
             mask = (indices == i)
             pred[mask] = np.polyval(self.coef_[i, :][::-1], x[mask])
 
-        bounded = (lb is not None or ub is not None)
-        if bounded:
+        if self.extrapolation == "constant":
+            pred[x < self._min_x] = y_min
+            pred[x > self._max_x] = y_max
+
+        elif self.extrapolation == "linear":
+            # Spline derivatives at the boundaries
+            deriv_min = 0
+            deriv_max = 0
+            for i in range(self.degree):
+                deriv_min += i * self.coef_[0, i]
+                deriv_max += i * self.coef_[-1, i]
+
+            mask = x < self._min_x
+            pred[mask] = y_min + (x[mask] - self._min_x) * deriv_min
+
+            mask = x > self._max_x
+            pred[mask] = y_max + (x[mask] - self._max_x) * deriv_max
+
+        if self.space == "log":
+            pred = np.exp(pred)
+
+        if self.extrapolation_bounds is not None:
             pred = np.clip(pred, lb, ub)
 
         return pred